Задачи за производство: code bigInt

+++ planning/TaskDetails.class.php
@@ -107,7 +107,7 @@ class planning_TaskDetails extends doc_Detail
     public function description()
     {
     	$this->FLD("taskId", 'key(mvc=planning_Tasks)', 'input=hidden,silent,mandatory,caption=Задача');
-    	$this->FLD('code', 'int', 'caption=Код,input=none');
+    	$this->FLD('code', 'bigint', 'caption=Код,input=none');
     	$this->FLD('operation', 'varchar', 'silent,caption=Операция,input=none,removeAndRefreshForm=code');
     	$this->FLD('quantity', 'double', 'caption=Количество,mandatory');
     	$this->FLD('weight', 'cat_type_Weight', 'caption=Тегло');
